PriceSort + buttons Rating working

+++ app/src/main/java/com/example/maartenvandenhof/studentmenu/Menu.java
@@ -20,7 +20,7 @@ public class Menu {
     private String description;
     private String recipe;
     private double price = 0;
-    private double rating;
+    private int rating;
     private ImageView imageToUpload;
 
 
@@ -116,11 +116,11 @@ public class Menu {
         this.recipe = recipe;
     }
 
-    public double getRating() {
+    public int getRating() {
         return rating;
     }
 
-    public void setRating(double rating) {
+    public void setRating(int rating) {
         this.rating = rating;
     }
